+++ library/ruler/src/main/java/io/mehow/ruler/LengthUnit.kt
@@ -2,8 +2,8 @@ package io.mehow.ruler
 
 import java.math.BigDecimal
 
-interface LengthUnit<T> : Comparable<T>, Iterable<T> where T : Enum<T>, T : LengthUnit<T> {
+interface LengthUnit<T> : Comparable<T>, Iterable<T> where T : LengthUnit<T>, T : Enum<T> {
   val meterRatio: BigDecimal
 
-  fun appliesRangeTo(meters: BigDecimal): Boolean
+  operator fun contains(meters: BigDecimal): Boolean
 }